update: menutoolbar - added nupp, exec task - kontrol

+++ models/libs/libraries/asutused.js
@@ -36,9 +36,9 @@ module.exports = {
         {
             sql: `SELECT (e.element ->> 'aa') :: VARCHAR(20)       AS aa,
                          $2 :: INTEGER                             AS userid,
-                         (e.element ->> 'kas_palk') :: BOOLEAN     AS kas_palk,
-                         (e.element ->> 'kas_raama') :: BOOLEAN    AS kas_raama,
-                         (e.element ->> 'kas_oppetasu') :: BOOLEAN AS kas_oppetasu,
+                         ((e.element ->> 'kas_palk') :: BOOLEAN)::INTEGER     AS kas_palk,
+                         ((e.element ->> 'kas_raama') :: BOOLEAN)::INTEGER    AS kas_raama,
+                         ((e.element ->> 'kas_oppetasu') :: BOOLEAN)::INTEGER AS kas_oppetasu,
                          row_number() OVER ()                      AS id,
                          libs.get_asutuse_aa(a.id, 'RAAMA'::TEXT)  AS default_aa
                   FROM libs.asutus a,
