+++ models/ou/rekv.js
@@ -85,11 +85,11 @@ module.exports = {
                          aa.pank,
                          aa.konto::VARCHAR(20),
                          aa.tp,
-                         kassa                                                     AS kassapank,
-                         $2                                                        AS userId,
-                         coalesce((aa.properties ->> 'kas_tulud')::BOOLEAN, FALSE) AS kas_tulud,
-                         coalesce((aa.properties ->> 'kas_kulud')::BOOLEAN, FALSE) AS kas_kulud,
-                         coalesce((aa.properties ->> 'kas_palk')::BOOLEAN, FALSE)  AS kas_palk
+                         kassa::INTEGER                                                     AS kassapank,
+                         $2                                                                 AS userId,
+                         coalesce((aa.properties ->> 'kas_tulud')::BOOLEAN, FALSE)::INTEGER AS kas_tulud,
+                         coalesce((aa.properties ->> 'kas_kulud')::BOOLEAN, FALSE)::INTEGER AS kas_kulud,
+                         coalesce((aa.properties ->> 'kas_palk')::BOOLEAN, FALSE)::INTEGER  AS kas_palk
                   FROM ou.Aa aa
                   WHERE Aa.parentid = $1`,
             query: null,
